+++ portal-ui/src/screens/Console/Permissions/AddPermission.tsx
@@ -520,9 +520,30 @@ class AddPermissionContent extends React.Component<
                       control={<Radio />}
                       label="Write Only"
                     />
+                    <FormControlLabel
+                        value="trace"
+                        control={<Radio />}
+                        label="Trace"
+                    />
                   </RadioGroup>
                 </FormControl>
               </Grid>
+              {action === 'trace' && (
+                  <React.Fragment>
+                    <Grid item xs={12}>
+                      <br />
+                    </Grid>
+                    <Grid item xs={12}>
+                      <Typography
+                          component="p"
+                          variant="body1"
+                          className={classes.errorBlock}
+                      >
+                        Trace displays tracing information for all buckets.
+                      </Typography>
+                    </Grid>
+                  </React.Fragment>
+              )}
               <Grid item xs={12}>
                 <br />
               </Grid>